backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
motor==3.5.0
pymongo==4.8.0
torch==2.4.0
torchvision==0.19.0
numpy==1.26.4
Pillow==10.4.0
EMD-signal==2.0.0
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0
httpx==0.27.2
aiofiles==24.1.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
websockets==12.0
